+++ src/types.ts
@@ -3,8 +3,6 @@ export interface SimpleOptions {
   variableLabel: string;
   variableMinimumThreshold: number;
   variableMaximumThreshold: number;
-  variableDefaultMinimumValue: number;
-  variableDefaultMaximumValue: number;
   rangePrefix: string;
   rangeSuffix: string;
   rangeDelimiter: string;